fix: transform add in function
